refactor: Utils에 index.js 추가
