+++ fe/src/Components/App.jsx
@@ -1,11 +1,12 @@
 import { useEffect, useState } from 'react';
 import styled from 'styled-components';
 
+import { fetchData } from '@/Utils';
+
 import { BestProducts } from '@/Components/BestProducts';
 import { CategoryProductsList } from '@/Components/CategoryProductsList';
 import { Header } from '@/Components/Header';
 import { OrderPanel } from '@/Components/OrderModal';
-import { fetchData } from '@/Utils/Utils';
 
 const Root = styled.div`
   width: 1440px;
